+++ apps/testapp/cmd/init.go
@@ -58,7 +58,7 @@ func InitCmd() *cobra.Command {
 				}
 			}
 
-			proposerAddress, err := rollcmd.CreateSigner(&cfg, homePath, passphrase)
+			proposerAddress, err := rollcmd.CreateSigner(cmd.Context(), &cfg, homePath, passphrase)
 			if err != nil {
 				return err
 			}