+++ types/utils_test.go
@@ -78,7 +78,7 @@ func TestGetFirstSignedHeader(t *testing.T) {
 			noopSigner, err := noop.NewNoopSigner(privKey)
 			assert.NoError(t, err)
 
-			firstSignedHeader, err := types.GetFirstSignedHeader(noopSigner, tc.chainID)
+			firstSignedHeader, err := types.GetFirstSignedHeader(t.Context(), noopSigner, tc.chainID)
 			assert.NoError(t, err)
 			assert.NotNil(t, firstSignedHeader)
 			assert.Equal(t, uint64(1), firstSignedHeader.Height())